// src/app/core/models/meeting-agenda.ts
import { AgendaPurpose } from './agenda-purpose';
import { BaseModel } from './baseModel';
import { Attachment } from './attachment';
import { MeetingVote } from './meeting-vote';
import { User } from './user';
import { UserComment } from './user-comment';
import { AgendaParticipant } from './agenda-participant';
import { AgendaPresenter } from './agenda-presenter';
export class MeetingAgenda extends BaseModel {
    id: number;
    agenda_title_ar: string;
    agenda_title_en: string;
    agenda_time_in_min: number;
    meeting_id: number;
    agenda_purpose_id: number;
    agenda_presenters: Array<number>;
    agenda_attachments: Array<Attachment>;
    files: Array<any>;
	attachments: Array<Attachment>;
	agenda_votes: Array<MeetingVote>;
	agenda_purpose: AgendaPurpose;
    presenters_agenda: Array<User>;
    agenda_user_comments: Array<UserComment>;
    canPresent?: boolean;
    agenda_description_ar: string;
    agenda_description_en: string;
    VotesNumber: number;
    current_user_comment: string;
    isShowMore: boolean;
	commentsMax: number;
	extraTime: number;
    is_presented_now: boolean;
    timer: number;
    agendaTempId: number;
    agenda_order: number;
    presenters: Array<AgendaPresenter>;
    participants: Array<AgendaParticipant>;
    can_access: boolean;
}